+++ src/solc/standard_json/output/contract/mod.rs
@@ -50,3 +50,21 @@ pub struct Contract {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub missing_libraries: Option<HashSet<String>>,
 }
+
+impl Contract {
+    ///
+    /// Checks if all fields are `None`.
+    ///
+    pub fn is_empty(&self) -> bool {
+        self.abi.is_none()
+            && self.metadata.is_none()
+            && self.devdoc.is_none()
+            && self.userdoc.is_none()
+            && self.storage_layout.is_none()
+            && self.evm.is_none()
+            && self.ir_optimized.is_none()
+            && self.hash.is_none()
+            && self.factory_dependencies.is_none()
+            && self.missing_libraries.is_none()
+    }
+}