+++ src/solc/standard_json/output/error/source_location.rs
@@ -39,13 +39,6 @@ impl SourceLocation {
     pub fn new_with_location(file: String, start: isize, end: isize) -> Self {
         Self { file, start, end }
     }
-
-    ///
-    /// Resolves the location, converting absolute offsets to line and column.
-    ///
-    pub fn resolve(&self) -> String {
-        self.file.to_owned()
-    }
 }
 
 impl FromStr for SourceLocation {